moved baseColors to colorsMetadata
