+++ packages/cli/src/colors/colorsMetadata.ts
@@ -1,0 +1,9 @@
+import type { CssColor, GlobalColors } from './types.js';
+
+export const baseColors: Record<GlobalColors, CssColor> = {
+  blue: '#0A71C0',
+  green: '#068718',
+  orange: '#EA9B1B',
+  purple: '#663299',
+  red: '#C01B1B',
+};
