+++ packages/cli/src/colors/index.ts
@@ -1,3 +1,4 @@
 export * from './utils.js';
 export * from './theme.js';
 export * from './types.js';
+export * from './colorsMetadata.js';
